Move DirFilenameDecoder to his own file
